Fixed a bug with angles - fixed a bug where the guard would not follow the hnter after a certain   angle

+++ include/Character.h
@@ -42,7 +42,7 @@ protected:
 	milliseconds _fireball_cooldown = milliseconds{500};
 	time_point _last_fireball_time = time_point::min();
 	std::uniform_int_distribution<> _fire_angle_error{0, 0};
-	int _fire_error_step = 1; // TODO: use this parameter
+	double _fire_error_step = 1; // nb of deg of precision lost for each 1/5 of health lost
 
 	Sound* _fire_sound;
 	Sound* _hit_sound;
